requirements.txt
aiofiles==0.8.0
asyncio==3.4.3
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dash==2.6.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==4.5
Flask==2.2.2
Flask-Compress==1.12
flask-sock==0.5.2
google-auth==1.35.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.1
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pandas-ta==0.3.14b0
patsy==0.5.2
plotly==5.10.0
priority==2.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
quart==0.18.0
requests==2.28.1
schedule==1.1.0
scipy==1.4.1
simple-websocket==0.8.0
six==1.16.0
soupsieve==2.3.2.post1
statsmodels==0.13.2
tenacity==8.0.1
toml==0.10.2
torchaudio==0.12.0+cu113
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
wsproto==1.2.0
zipp==3.8.1
zope.interface==5.4.0
